feat: move docker cleanup to its own tab
